Added follow and unfollow commands to a fan.
These will add the leader_board to the active record object as well as add it to the set in redis for that fan.  A leaderboard will also have a set of fans that are following it.

added a redis_key function to each of the models.  This is a helper that will create a prefix for the object to make a key.  The passed in string is appended to name what the key is for.

For example leader_board:1:followers is used for the set of followers following the leader_board with an id of 1 in the sql datastore.

Added a top_10_scores method to the leader_board.  This will get the highest 10 scores stored, returned in decending order by points.

The regular scores method will return all scores in the database.